requirements.txt
uvloop>=0.19.0
websockets>=12.0
orjson>=3.9.0
python-dotenv>=1.0.0
numpy>=1.26.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
ccxt>=4.0.0
tqdm>=4.66.0
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
flet>=0.21.0
scikit-learn>=1.3.0
